requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
cloudinary==1.41.0
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.16
gunicorn==23.0.0
idna==3.10
packaging==24.1
psycopg2==2.9.10
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
django-allauth==65.0.2
whitenoise==6.7.0
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
PyJWT==2.9.0
cryptography==43.0.3
cffi==1.17.1
pycparser==2.22
djangorestframework==3.15.2
django-cors-headers==4.5.0
drf-yasg==1.21.8
inflection==0.5.1
pytz==2024.2
PyYAML==6.0.2
uritemplate==4.1.1
dj-rest-auth==6.0.0
